chore(api/package.json): Remove unused lib from api
A transitive dependency of that package (node-webcrypto-ossl)
was breaking the build on my machine. Since it's unused, we
could simply remove it.

Possibly fixes the issues on other Macs too, but idk. Would be
good to know if anyone was managing to use it on a Mac to know
if something is wrong on my machine as well.