Ignore test and test data files from babel
This was one of the ways of fixing the jest issue,
to avoid having test files under the dist package.

It did not fix it completely because we still have
a `test.js` file on the root of the package (lmk if
we should remove it too)

It also doesn't fix the root cause of the issue,
which I found later.