Create unified  stream patch API with push-target
I found 2 different PATCH APIs in the /stream endpoint,
each of them changing separate fields. I would be
creating a 3rd one for just another field, but decided
it was interesting to have a single patch instead which
can be used for changing any of the mutable fields
depending on the payload.

So I did that. And also ended up making our schemas
support references to one another since the pain was
too big to implement all the schema validation by hand
or to repeat once more all those push-target fields.