Move ACL to api-token
Still keep retro-compatibility both with:
 - existing api-tokens without an access field by defaulting
   their permissions to "all";
 - existing APIs with no access retriction, by defaulting their
   access requirements to "all".

So:
 - existing tokens keep working on any API since they have access to
   everything, even if the APIs evolve to start requiring something
   like "writeAccess"
 - existing APIs can migrate to the access list either a specific or a
   very broad one, so that new tokens can be created with granular
   access.
 - non-token access is still able to access everything since there will
   be no token object, so worst case it just defaults to "all"

Finally, the broadcaster hooks require both a token with the "broadcasterHooks"
permission but also coming from an account/userId that is an admin. So only we
will be able to create an api-token with access to those and also it
will be granular enough giving access to no other authenticated API,
since all the others will default to requiring an "all" permission which
the broadcaster token won't have.