feat(api): Allow restricting access for API tokens via resource/method rules (#483)
* Create user broadcaster flag for go-livepeer access

* Make it a bit less generic

* Move access validation to auth middleware

* Move ACL to api-token

Still keep retro-compatibility both with:
 - existing api-tokens without an access field by defaulting
   their permissions to "all";
 - existing APIs with no access retriction, by defaulting their
   access requirements to "all".

So:
 - existing tokens keep working on any API since they have access to
   everything, even if the APIs evolve to start requiring something
   like "writeAccess"
 - existing APIs can migrate to the access list either a specific or a
   very broad one, so that new tokens can be created with granular
   access.
 - non-token access is still able to access everything since there will
   be no token object, so worst case it just defaults to "all"

Finally, the broadcaster hooks require both a token with the "broadcasterHooks"
permission but also coming from an account/userId that is an admin. So only we
will be able to create an api-token with access to those and also it
will be granular enough giving access to no other authenticated API,
since all the others will default to requiring an "all" permission which
the broadcaster token won't have.

* Make auth middleware typescript

Easy one, but same :satisfaction:

* Make authorization a "some" instead of an "every"

Requiring a credential to have "any" specific access flag
sounds more useful than requiring "every" one to be present.

e.g. a "read" operation on a stream could require either a
"readStream" or a "readAll" access flag.

* Make the ACLs matching logic generic

They can now either be:
 - a single string with the required access
 - an object with an "any" field with a list of access items which, if
   any are present, will provide access to the resource
 - an object with an "all" field with a list of access items that all
   need to be present.

And the object is actually recursive, so an "all" array can have an item
which is an object with an "any" field and vice versa.

a

* Example: Make "admin" be an access item as well

* Revert "Example: Make "admin" be an access item as well"

This reverts commit ad9b6b340526f7cb562d521ba5bba3e1b2e7c43e.

* Move to a role-based authorization model

Using http-hash for route matching (second choice was
https://www.npmjs.com/package/url-router but it had much less downloads)
and keeping policies inline in the code for simplicity.

* Only call isAuthorized if policies exist in token

* Inline auth rules in api-tokens

* Implement prefix trimming and rules validations

Both on auth middleware and on api-token creation API.

* Move setupUsers to test-helpers

Will facilitate creating new tests like this one for the auth
middleware.

* Set up user-creation boilerplate for auth.test.ts

* Create separate errorHandler middleware

To be reused on tests and all
Also improved types overall by forcing the explicit
Express type and created a helper isAPIError function

* Attempt on magic typing of TestServer

Didn't really work so I will just use
db static directly. I think the changes
are worth keeping though.

* Create initial real tests for auth middleware

Testing base cases having no rules listed or
having especifically an empty rules list.

* Create tons of tests for access rules

* Add tests on api token creation rules validation

* Revert changes to stream controller

* Add some corner case tests to http prefix

* Disallow api-token API access with api-tokens

Please enable ignore whitespace when reviewing this

* Fix auth for api groups/routers

* Join baseUrl instead of splitting originalUrl