api/webhooks: Only ack/nack webhook event once
With the try/catch block inside the triggers.map
transformation, it meant we could nack the event
multiple times, once for each webhook that failed.

Additionally, the `return` in there didn't mean much,
since we would only return from the inner lambda but
not from the outer handleEventsQueue function.

Moved try/catch to outside the Promise.all, so that
we only nack once and make sure to return from
handleEventsQueue and avoid the `ack` in the end of
the function.