www: POC of enable/disable switch in targets
It's currently enabling/disabling recording,
but we only need to create one that actually
disables the push targets now which should be
pretty easy

Had way too many issues with the useQuery hooks...
I had working version without the changes to Table,
if preferred. But that version used like a useRef
to the real "fetcher" function, so that it picked
the right closure on every execution. Felt more like
a hack than the current one, so I'm keeping this
one unless someone prefers the useRef.