api: Add .env.local file to keep stripe API off git
It's a bit annoying to keep having to skip the changes to
the .env file not to commit sensitive information to git.

With .env.local, we can completely ignore that file from git
and add only the sensitive info there, while loading the
regular .env file for whatever is not present in the local
one.