api: Implement common protocol for RabbitMQ messages (#569)
* api: Improve resp info on webhook_response

We were not serializing much about the repsonse
other than the statusCode, since the native
`Response` object from fetch doesn't serialize
nicely as JSON. So we have to extract some of
the properties that we want from it.

* api: Also store how long it took to call a webhook

* api/webhooks: Create separate message types ns/file

* api/webhooks: Create separate WebhookTrigger msg

This also fixes retries, which actually seemed to be broken.
They sent an "event" is if it were a trigger for retry, and
most probably err'd on the second try.

* api/store: Add some type safety to rabbit-queue

Now we it will be easier to guarantee we follow the
global RabbitMQ protcol being defined between livepeer
infra services.

Messages should always have a `type` key specifying what's
the type of the object so it can be deserialized properly.
Also made sure webhook event routing keys always include
the event itself, so one can consistently grab all stream,
multistream, recording events, etc.

* api/webhooks: Only ack/nack webhook event once

With the try/catch block inside the triggers.map
transformation, it meant we could nack the event
multiple times, once for each webhook that failed.

Additionally, the `return` in there didn't mean much,
since we would only return from the inner lambda but
not from the outer handleEventsQueue function.

Moved try/catch to outside the Promise.all, so that
we only nack once and make sure to return from
handleEventsQueue and avoid the `ack` in the end of
the function.

* api/store: Implement RabbitMQ livepeer protocol

* api/webhooks: Move sessionId field out of payload

Since the information is not meant to the user but for our
own cannon logic, we can make it a first party field in the
RabbitMQ message. IMO ideally, the payload should contain
only information that should be sent on the webhook call,
not for our internal use.

* api/webhooks: Fix tests (among other things)

Also fix the new event queue, to avoid having the
same binding as before getting all events in the
default queue.

Also fix delayed messages emitting. We were
both accumulating duplicate setup functions as
well as re-using the same queues for different
routing keys.

* api/webhooks: Create test for webhook retries

* api/webhooks: Create helper processWebhookEvent

I broke the tests initially because I forgot to ack
the message. Since it seems very error prone to require
individual ack/nacks like that, created a helper just
like for trigger processing, ack/nacking the message on
the finally block.

* Fix tests

forgot a final return on the helper created on previous commit

* api/webhooks: Log bad event instead of throwing error

This will avoid the infinite retry loop on a bad event,
but still avoid keeping us blind in case bad events start
lurking into the internal queue. That must never happen.

* s/semaphs/sems/g

* api/stream: manifestID=>streamID on rabbit events