www: Multistream frontend  (#579)
* www: Create getMultistreamTarget function in API client

* www: Create simplest MultistreamTargetsTable

 - Copied from SessionsTable
 - Made all rows text for simplicity
 - No current status
 - No way to enable/disable targets
 - No way to add new targets

* www: Add missing webhook events to WebhookDialog

* www: POC of enable/disable switch in targets

It's currently enabling/disabling recording,
but we only need to create one that actually
disables the push targets now which should be
pretty easy

Had way too many issues with the useQuery hooks...
I had working version without the changes to Table,
if preferred. But that version used like a useRef
to the real "fetcher" function, so that it picked
the right closure on every execution. Felt more like
a hack than the current one, so I'm keeping this
one unless someone prefers the useRef.

* www: Fix build

* www: Implement actual toggling of multistream targets

* www: Fetch each target on its separate query

* www: POC of overflow menu with delete option

Design/layout is still horrible, but I'll work on the
functionality first and hope that the layout is orthogonal
to all of that, easily tunable after functionality is done
(especially if pairing with some front-end master 👀

Also created the overflow with just one option for now, to
give room to adding more stuff (like "Edit") soon.

* api: Add .env.local file to keep stripe API off git

It's a bit annoying to keep having to skip the changes to
the .env file not to commit sensitive information to git.

With .env.local, we can completely ignore that file from git
and add only the sensitive info there, while loading the
regular .env file for whatever is not present in the local
one.

* www: Implement multistream target deletion for real

* www: Create simple create target dialog

Currently only supports changing the URL, which
automatically is created a name on the server
based on the URL host name. Profile hardcoded to
be always source.

Next steps are to allow modifying the profile
and the name.

Then finally grabbing multistream state from
livepeer-data/analyzer. We're close!

* www: Multi-polishing

* www: Fix toolbox file name

* www: Finish target creation dialog

* www: Implement multistream target status badge

Did I hear feature completeness?! 🙌

* www: Small optimization to status fetching

Avoid fetching when there are no targets we are
interested in.

* Some final small fixes

Bedtime 🥱

* www: Fix create&delete dialog lifecycles

 - Delete dialog not getting the screen stuck anymore after closing
 - Create dialog not maintaining the data even after submitting
   successfully

* www: Fix reading of connected status

* www: Refresh stream details after 5 seconds

This will make the Idle=>Online detection automatic.

Also included some other changes:
 - Move health fetching to page, just besides the stream, since it is a
   global information that we might want to start using in other
   components pretty soon (like for the global stream Idle indicator or
   the stream health page itself)
 - Make AnalyzerClient receive the region as an argument. Will be less
   error-prone when creating the query keys, since it's quite easy to
   forget about a key that is not sent as an argument in the query (and
   we cant use functions as query keys, as I've tested).

* www: ID tooltip and couple updates to dialog msgs

* www/stream/[id]: Fix font for deletion dialogs

* www: Add tooltip with time ago to status badges