[POC] www/multistream: Consider session for status badge
This is not fully working because the session response
apparently takes a while to update so we end up displaying
invalid states anyway as if we were on the previous stream
(like all targets offline or something). The true/correct
version for this will need changing the analyzer logic to
actually consider sessions or maybe keep an "active" state
for the streams.

I'll revert the change, but pushing to GitHub for historic
purposes.