api/package.json: Fix local docker build command
We were forcing to provide a VERSION env when building locally
as well due to an accidental revert of a revert on a revert of
another change. This fixes that by just moving the --help
sanity check to inside the docker build instead of in the
docker build command, which is just simpler as well.

With this, we don't even need the intermediate TAGS var either
since the DOCKER_TAGS is only used once.