.github: Create separate api CI workflow
Building the API image takes a looong time today
because we build all the WWW stuff before it.

This separates the API docker image build so that
we can have it in a few minutes, not in 10 minutes.

The downside is that we might still publish images
that don't pass the yarn tests. We could include the
yarn test steps on the workflow to avoid that, but
that will add 1-2 minutes to it as well. I'll look
into making it in parallel perhaps?