.github/api: Ignore 'latest' branch for API build
I give up on trying to make the build from the latest
branch have the exact same sha256 file. I think it's
just not meant to be like that as Docker will use all
kinds of metadata on the sha and end up with different
hashes anyway.

We can just ignore it instead and the latest builds won't
mess up with the tags in our livepeerci/api images.