.github: Create separate api CI workflow (#817)
* .github: Create separate api CI workflow

Building the API image takes a looong time today
because we build all the WWW stuff before it.

This separates the API docker image build so that
we can have it in a few minutes, not in 10 minutes.

The downside is that we might still publish images
that don't pass the yarn tests. We could include the
yarn test steps on the workflow to avoid that, but
that will add 1-2 minutes to it as well. I'll look
into making it in parallel perhaps?

* .github: Add install and test steps to API CI

* .github: Add services to API CI so tests pass

* .gthub: Make sure to run in packages/api folder

* Revert ".gthub: Make sure to run in packages/api folder"

This reverts commit a57127c89f488e7b72dca92c7a5dc378723528e6.

* .github: Add yarn dep caching on API build

* prettier lol

* api/package.json: Remove --cache-from in docker build

I suspect the download is taking more than the advantage it brings

* api: crazy-test: parallelize test and docker build

* Revert "api: crazy-test: parallelize test and docker build"

This reverts commit 6656f0edef6d68b87b72281478949a7f423a322f.

* api/dockerfile: Add build ARGs only in the last minute

* .github/api: Ignore 'latest' branch for API build

I give up on trying to make the build from the latest
branch have the exact same sha256 file. I think it's
just not meant to be like that as Docker will use all
kinds of metadata on the sha and end up with different
hashes anyway.

We can just ignore it instead and the latest builds won't
mess up with the tags in our livepeerci/api images.