api: Call Sendgrid email validation API on signup (#850)
* api: Create api helper to validate an email address

* api: Call validate email func on signup

* api: Fix typo

* api: Convert helpers to TypeScript

* api: Separate async email validation function

Cleaner than the .then.catch IMO

* api: Receive separate arg for validating email addresses

I wouldn't do that if Sendgrid didn't obly us to do that :sweat:

They don't support API keys that can both send emails and validate
addresses. I know wtf.

* api: Add back isomorphic-webcrypto

* api/helpers: Remove unused imports

* api/users: Add FieldsMap type

* api: Fix tests

In the end it was a problem with our mock which wasn't
following the interface. Decided to just revert the old
change to be compatible instead regarding the personalization.to
but changed the snakes to camels instead.

* api: Do not call validation API if no API key provided