api/queue: Reduce the number of delayed queues
Removing the routing key from the queue name we
avoid creating so many queues. For that, required
creating a separate exchange for the queue as well
so that the original routing key of the message is
kept in the message metadata, and forwarded to the
deadletter exchange when the message expires.