api: Allow hackers to create dev API keys with CORS (#985)
* api: Split authenticator middleware from the authorizer

* api: Make createdByTokenId an admin-only field

That field basically exposes API keys on all stream
API endpoints. Which means for example that any API
key with access to the /streams API can get access
to any other API key that has ever created streams
(since our API keys are actually the IDs themselves).

This avoids that leak for non-admin users. We should
eventually make the API keys themselves not actually
use the ID as the secret key.

* api: Create separate 401 error different from 403

401 actually means the request is not authenticated (yeah the HTTP name is
    confusing)
403 means that the caller is authenticated but is not allowed to perform
    the action (aka forbidden or unauthorized)

* api: Move CORS and Prometheus to app router

We'll need to add middlewares before CORS now, then
I also didn't want to leave Prometheus all alone in
the root (that test middleware doesnt count).

* api: Validate noApiToken option

* api: Allow API keys to add allowed origins

* api: Add Vary: Authorization to all responses

* api: Fix tests and auth middleware imports

* api: Remove CORS logic from asset controller

* api/cli: Make cors allow-list configured as an arg/env

* api/cors: Make the static allowlist only for JWT reqs

* api/env: Remove localhost from default allowlist

We should still configure prod/staging separately, with
prod only really allowing access from livepeer.com (and maybe explorer?)

* api: Remove explorer from CORS allowlist

Not needed anymore!

Also fix some . that should be \. in the regexes

* api: Default CORS to https://${frontendDomain} instead

* api: Create logic for restricting CORS access

Pending actually allowing some APIs to be used

* api: Allow some stream APIs to use use CORS

* api: Allow some multistream APIs to use CORS

* api: Allow some asset APIs to use CORS

* api: Allow any origin to access /api/asset/upload

* www: Create allow CORS checkbox in API key creation

* www: Add input for allowed origins for API key CORS

* www: Add a + button next to add origin field

* www: Clear API key dialog anytime it is closed

* www: Update copy for less indirection

* www: Allow creating CORS API keys with full API access

* prettier

* www: Fix default suggested CORS origin

* www: Stop disallowing whitespace in token

* www: Remove whitespace from addAllowedOrigin field

* api: Allow any origin in the OPTIONS call

* api: Create getCorsOpts internal helper

Just some cleaner code

* api: Disallow API keys on certain user APIs

Also disallow an user to suspend themelves lol

* api: Fix CORS for API tokens that allow *

* api: Simplify OPTIONS CORS handling

* api: Fix CORS allow-list arg parsing

* api: Make CORS allowed origins list a JSON array

* api/task: Tune authorizer params

* api: Always run cors middleware after auth

* api: Fix auth middleware

express-async-errors is a lie, we can't really
ignore next for a middleware apparently.

* www: Create column for CORS in API key table

* www: Add origin validation

* www: Keep only * when * is in allowed origins

* www/docs: Add first part of API key CORS docs

* www: Add dummy API access docs and links from creation dialog

* api: Change transcode API to /api/asset/:id/transcode

Still keep backward compatibility since the NFT SDK is already
calling the old one.

* www/docs: Fix API key reference

Made it a little shorter just with usage information
and a link to the getting an API key guide.

* www/docs: Remove domain warning from Stream Health

It is not true anymore!

Also remove the repeated beta badge that messes with the
header. We already have the note right below and it's not
**that** beta anymore lol

* www/docs: Finish API key docs!

* www/docs: Final nits!