api: Avoid suspending user streams when suspending the user
It is enough to just suspend the user, which will already make
all streams effectively suspended, but only terminate the streams
instead of marking them as suspended as well.

This allows us to potentially revert the suspension more easily,
by not changing the state of every stream in the user account.