api: Fix cross-OS direct uploads
We will be uploading files to the old default bucket (which can become
only for direct uploads), while the final files processed by catalyst
will go to a different bucket.

To fix, send the public URL of the file for processing instead of the
uploaded object key. That's what's gonna reach catalyst anyway.