api: Stop setting isActive from stream-info-app
It is leading to duplicate assets generated as oftentimes
stream-info-app sets isActive to true again after the
recording asset had already been generated. Instead of that,
just create every child session as active and treat the
isActive as a flag for whether the recording has been generated
or not (without a chance it is made true again).