api: Add createdByTokenName field to assets (#2034)
* api: Simplify validateAssetPayload interface

We need a lot of stuff from the request, and
now something else (req.token). So let's just
send the whole req for simplicity.

* api: Add createdByTokenName field to asset

* api: Stop saving API key on every stream object

The API key ID is basically the ID itself. Let's
stop copying it around so much. This is not really
used anywhere but to leak tokens lol