Reduce timeout for auto stream start on pull API
The stream pull API was proving too slow so I'm reducing the auto-start timeout significantly, the processing for that request should still continue even when dropping it from the studio side (needs more testing). We then need to catch these timeout errors in a try catch and not propagate these to the user. So this auto start becomes more of a best effort process which I think is acceptable since even if something goes wrong the stream startup will happen when viewers start making requests.